+++ src/Renderer/RenderTexture.cpp
@@ -56,7 +56,9 @@ namespace uth
             }
 
             uth::Graphics::BindRenderBuffer(m_depthBuffer);
-            uth::Graphics::SetRenderBufferStorage(size.x, size.y);
+            uth::Graphics::SetRenderBufferStorage(
+				static_cast<unsigned int>(size.x), 
+				static_cast<unsigned int>(size.y));
             uth::Graphics::BindRenderBufferToFrameBuffer(m_frameBuffer, m_depthBuffer);
         }
